Refactoring the unit test file and adding more tests
